+++ pkg/store/client.go
@@ -14,7 +14,7 @@ type Client interface {
 	Get(path string) (io.ReadCloser, error)
 
 	// Put saves a file to the store
-	Put(path string, r io.Reader, sz int64) (int64, error)
+	Put(path string, r io.Reader, sz int32) (int32, error)
 
 	// Remove deletes an object from the store
 	Remove(path string) error
